requirements.txt
#django
#gunicorn
#django-heroku
requests-html
arcgis
#pandas
numpy
matplotlib
#cartopy
#watermark
#pprint
mpld3
scipy
#psycopg2-binary
psycopg2
dj-database-url
python-dotenv
sklearn
numdifftools
#lmfit
statsmodels
#numba
memory_profiler
line_profiler
apscheduler
-e git://github.com/prediksicovidjatim/lmfit-py.git#egg=lmfit-py
-e git://github.com/R-N/prediksicovidjatim-core.git#egg=prediksicovidjatim-core